* use u16 for col and row

+++ src/components/board.rs
@@ -5,16 +5,16 @@ use yew::{html, Component, Context, ContextProvider, Html, Properties};
 
 #[derive(Properties, PartialEq, Clone)]
 pub struct BoardProps {
-    pub col: u32,
-    pub row: u32,
+    pub col: u16,
+    pub row: u16,
 }
 
 pub enum BoardMsg {
     LockVal((u16, CellValue)),
 }
 pub struct Board {
-    pub col: u32,
-    pub row: u32,
+    pub col: u16,
+    pub row: u16,
     state:Rc<BoardState>
 
 }
@@ -55,7 +55,7 @@ impl Component for Board {
     fn view(&self, ctx: &Context<Self>) -> Html {
         let state = self.state.clone();
         let onselect = ctx.link().callback(BoardMsg::LockVal);
-        let cells = (0..self.col*self.row).map(|i| html! { <Cell id={i as u16} onselect={onselect.clone()}/>  });
+        let cells = (0..self.col*self.row).map(|i| html! { <Cell id={i} onselect={onselect.clone()}/>  });
         let style = format!("grid-template-columns: repeat({}, 1fr);grid-template-rows: repeat({}, 1fr);", self.col, self.row);
         html! {
             <ContextProvider<Rc<BoardState>> context={state}>
